integration tests for delete accounts

+++ src/functions/deleteAccounts/handler.js
@@ -1,4 +1,5 @@
 import Accounts from '../../database/accounts';
+import Customers from '../../database/customers';
 import generateResponse from '../../lib/generateResponse';
 
 export const handler = async (event) => {
@@ -6,10 +7,21 @@ export const handler = async (event) => {
   const { customerId, accountId } = event.pathParameters || {};
 
   const accounts = Accounts(userId);
+  const customers = Customers(userId);
   try {
     if (accountId) {
-      await accounts.deleteAccount(accountId);
+      const deletedItem = await accounts.deleteAccount(accountId);
+      if (!deletedItem) {
+        return generateResponse({
+          statusCode: 404,
+        });
+      }
     } else {
+      if (!(await customers.getCustomer(customerId))) {
+        return generateResponse({
+          statusCode: 404,
+        });
+      }
       await accounts.deleteAllAccounts(customerId);
     }
